backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
openai==1.58.1
anthropic==0.42.0
google-genai==1.5.0
httpx==0.28.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
duckduckgo-search==7.2.1
markdown==3.7
aiofiles==24.1.0
